requirements.txt
NetflixRouletteAPI==5.0
requests==2.7.0
